pyproject.toml
[project]
name = "navya-local-dev-server"
version = "0.1.0"
description = "2026 High-Performance Local Testing Mesh"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "torch>=2.2.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.12.0",
    "pandas>=3.0.2",
    "pygithub>=2.9.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
